Added Java byString and byLog10
 On branch main
 Your branch is up to date with 'origin/main'.

 Changes to be committed:
	new file:   1000integers.bin
	new file:   1000integers.txt
	new file:   cLangTest.o
	modified:   cTest.h
	new file:   cTest.o
	modified:   logTest.c
	modified:   logTest.o
	modified:   notes.txt
	modified:   rndIntegers.c
	modified:   rndIntegers.o
	new file:   rustyLogTest.o
	new file:   rustyString.o
	modified:   rustyString.rs
	new file:   stringTest.class
	new file:   stringTest.java

+++ stringaling/cTest.h
@@ -6,7 +6,7 @@ int numDigits(unsigned int i)
         return 1;
     int count = 0;
     while (i != 0) {
-        i = i / 10;
+        i =(unsigned int) i / 10;
         ++count;
     }
     return count;
